Refactoring and spawners that work
