+++ src/gnat/attack.py
@@ -1,0 +1,4 @@
+from src.gnat.gnat import Gnat
+
+class Attack(Gnat):
+    ...
